+++ src/main/java/org/isec/pa/ecossistema/Fauna.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.isec.pa.ecossistema;
 
 public class Fauna {
