+++ src/main/java/org/isec/pa/ecossistema/model/data/EcossistemaData.java
@@ -1,2 +1,4 @@
-package org.isec.pa.ecossistema.model;public class EcossistemaData {
+package org.isec.pa.ecossistema.model.data;
+
+public class EcossistemaData {
 }
