+++ src/main/java/org/isec/pa/ecossistema/model/data/FaunaData.java
@@ -1,4 +1,4 @@
 package org.isec.pa.ecossistema.model.data;
 
-public class InanimadosData extends EcossistemaData{
+public class FaunaData extends EcossistemaData{
 }
