+++ src/main/java/org/isec/pa/ecossistema/model/data/FloraData.java
@@ -1,4 +1,4 @@
 package org.isec.pa.ecossistema.model.data;
 
-public class FaunaData extends EcossistemaData{
+public class FloraData extends EcossistemaData{
 }
