+++ src/main/java/org/isec/pa/ecossistema/model/data/InanimadosData.java
@@ -1,2 +1,4 @@
-package org.isec.pa.ecossistema.model.data;public class InanimadosData {
+package org.isec.pa.ecossistema.model.data;
+
+public class InanimadosData extends EcossistemaData{
 }
