+++ src/main/java/org/isec/pa/ecossistema/model/fsm/FaunaState.java
@@ -1,4 +1,4 @@
 package org.isec.pa.ecossistema.model.fsm;
 
-public class FaunaContext {
+public enum FaunaState {
 }
