+++ src/main/java/org/isec/pa/ecossistema/model/fsm/FaunaStateAdapter.java
@@ -1,2 +1,6 @@
-package org.isec.pa.ecossistema.model.fsm;public class FaunaStateAdapter {
+package org.isec.pa.ecossistema.model.fsm;
+
+import org.isec.pa.ecossistema.model.fsm.Faunastates.IFaunaState;
+
+public abstract class FaunaStateAdapter implements IFaunaState {
 }
