+++ src/main/java/org/isec/pa/ecossistema/model/fsm/Faunastates/EatingState.java
@@ -1,4 +1,4 @@
 package org.isec.pa.ecossistema.model.fsm.Faunastates;
 
-public class InitialState {
+public class EatingState {
 }
