+++ src/main/java/org/isec/pa/ecossistema/model/fsm/Faunastates/IFaunaState.java
@@ -1,2 +1,4 @@
-package org.isec.pa.ecossistema.model.fsm.Faunastates;public interface IFaunaState {
+package org.isec.pa.ecossistema.model.fsm.Faunastates;
+
+public interface IFaunaState {
 }
