+++ src/main/java/org/isec/pa/ecossistema/model/fsm/Faunastates/InitialState.java
@@ -1,2 +1,4 @@
-package org.isec.pa.ecossistema.model.fsm.Faunastates;public class InitialState {
+package org.isec.pa.ecossistema.model.fsm.Faunastates;
+
+public class InitialState {
 }
