+++ src/main/java/org/isec/pa/ecossistema/model/fsm/Faunastates/ReproducingState.java
@@ -1,4 +1,4 @@
 package org.isec.pa.ecossistema.model.fsm.Faunastates;
 
-public class EatingState {
+public class ReproducingState {
 }
