+++ src/main/java/org/isec/pa/ecossistema/ui/EcossistemaUI.java
@@ -1,2 +1,4 @@
-package org.isec.pa.ecossistema.ui;public class EcossistemaUI {
+package org.isec.pa.ecossistema.ui;
+
+public class EcossistemaUI {
 }
